CCC-7 Move gateway and CertContractService to gateway pkg
