+++ cert-app/fabric-gateway/gateway/gateway.go
@@ -1,4 +1,4 @@
-package main
+package gateway
 
 import (
 	"crypto/x509"
